requirements.txt
black
flake8
flake8-pyproject
pillow